Add files via upload
Finished insertion code and added some sample data for location, user, and review. DB should build locally with no problems.

+++ db_src/create_tables.py
@@ -45,7 +45,7 @@ CREATE TABLE review (
     review_id INT NOT NULL AUTO_INCREMENT,
     user_id INT NOT NULL,
     location_id INT NOT NULL,
-    stars DECIMAL NOT NULL,
+    stars FLOAT NOT NULL,
     comment VARCHAR(256),
     review_image LONGBLOB,
     PRIMARY KEY (review_id)
